Revert 199575 "Fix more 64 bit compile errors flagged by convers..."
> Fix more 64 bit compile errors flagged by conversions from size_t to uint32 and to int.
> Looks like these were masked in webkit_glue.gypi.
> 
> Using static_cast for now.
> 
> BUG=237249
> TBR=jam
> 
> Review URL: https://codereview.chromium.org/14743008

TBR=ananta@chromium.org

Review URL: https://codereview.chromium.org/15038004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@199589 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/plugins/npapi/plugin_stream_win.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2013 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -27,8 +27,7 @@ void PluginStream::WriteAsFile() {
 size_t PluginStream::WriteBytes(const char *buf, size_t length) {
   DWORD bytes;
 
-  if (!WriteFile(temp_file_handle_, buf, static_cast<uint32_t>(length),
-                 &bytes, 0))
+  if (!WriteFile(temp_file_handle_, buf, length, &bytes, 0))
     return 0U;
 
   return static_cast<size_t>(bytes);
